refactored CertificationReportPage.jsx
