+++ web-ui/src/components/certifications/CertificationsTable.css
@@ -1,0 +1,38 @@
+#certification-report-page {
+  margin: 0 auto;
+  padding: 1rem;
+
+  /* .add-button {
+    border: 1px solid red;
+  } */
+
+  .column {
+    display: inline-flex;
+    flex-direction: column;
+    align-items: flex-end;
+  }
+
+  table {
+    border-collapse: collapse;
+
+    img {
+      max-height: 5rem;
+    }
+  }
+
+  td,
+  th {
+    border: 1px solid gray;
+    padding: 0.5rem;
+  }
+
+  th {
+    background-color: var(--oci-orange);
+  }
+}
+
+.certification-report-dialog {
+  .MuiFormControl-root {
+    margin-top: 2rem;
+  }
+}
